#22 REFACTOR. AlbumRes 패키지 위치 변경
