+++ src/main/java/com/sparta/topster/domain/maniadb/service/ManiadbService.java
@@ -1,9 +1,7 @@
 package com.sparta.topster.domain.maniadb.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.sparta.topster.domain.album.dto.AlbumRes;
-import com.sparta.topster.domain.album.entity.Album;
-import org.w3c.dom.stylesheets.LinkStyle;
+import com.sparta.topster.domain.album.dto.res.AlbumRes;
 
 import java.util.List;
